// src/commands/configuration-management/interfaces/batch-export-import.constants.ts
export enum BatchExportImportConstants {
  STUDIO_FILE_NAME = "studio.json",
  VARIABLES_FILE_NAME = "variables.json",
  MANIFEST_FILE_NAME = "manifest.json",
  STUDIO = "STUDIO",
  APP_MODE_VIEWER = "VIEWER",
  ZIP_EXTENSION = ".zip",
  JSON_EXTENSION = ".json",
  NODES_FOLDER_NAME = "nodes/",
  SCENARIO_NODE = "SCENARIO",
}